+++ js/color-picker.js
@@ -14,25 +14,4 @@ ColorPicker.prototype.addColorUI = function(color) {
   // 色を js から増やせたりしたいが、html を弄るので後で。
   // その時は、追加したNodeにListenerをつけるのを忘れない！
 };
-
-function LineWidthPicker(elem, defaultLineWidth) {
-  // このelem には、input[type="range"] 要素が入ってくる（はず）。
-  this.element = elem;
-  this.lineWidth = defaultLineWidth;
-  this.element.addEventListener("change", event => {
-    this.lineWidth = event.target.value;
-  });
-}
-
-LineWidthPicker.prototype.changeMaxLineWidth = function(maxLineWidth) {
-  // TODO
-};
-
-LineWidthPicker.prototype.changeMinLineWidth = function(minLineWidth) {
-  // TODO
-};
-
-module.exports = {
-  ColorPicker: ColorPicker,
-  LineWidthPicker: LineWidthPicker
-};
+module.exports = ColorPicker;
